Module Octez_crawler.Layer1_event

This module defines functions that emit the events used by the layer 1 chain (see Layer_1).

val starting : name:string -> unit Lwt.t

val stopping : name:string -> unit Lwt.t

val connection_lost : name:string -> unit Lwt.t

Emits the event that the connection to the Tezos node has been lost.

val connection_timeout : name:string -> timeout:float -> unit Lwt.t

Emits the event that the connection to the Tezos node has timeouted.

val cannot_connect : name:string -> count:int -> Tezos_base.TzPervasives.tztrace -> unit Lwt.t

cannot_connect ~count error emits the event that the rollup node cannot connect to the Tezos node because of error for the count's time.

val wait_reconnect : name:string -> float -> unit Lwt.t

wait_reconnect delay emits the event that the rollup will wait delay seconds before attempting to reconnect to the Tezos node .

val switched_new_head : name:string -> Tezos_base.TzPervasives.Block_hash.t -> int32 -> unit Lwt.t

switched_new_head hash level emits the event that the layer 1 has notified a new head with hash at some given level.

val connected : name:string -> unit Lwt.t

val stopping_old_connection : name:string -> unit Lwt.t

val reconnect_connecting : name:string -> unit Lwt.t

val reconnect_notified : name:string -> unit Lwt.t

val reconnect_disconnected : name:string -> unit Lwt.t

val reconnect_connected : name:string -> unit Lwt.t