Module Octez_crawler.Reorg

type 'block t = {

old_chain : 'block list;
(*
The blocks that were in the old chain and which are not in the new one.
*)

new_chain : 'block list;
(*
The blocks that are now in the new chain. The length of old_chain and new_chain may be different.
*)

}

Type of chain reorganizations.

val no_reorg : 'a t

val encoding : 'a Tezos_base.TzPervasives.Data_encoding.t -> 'a t Tezos_base.TzPervasives.Data_encoding.t

val map : ('a -> 'b) -> 'a t -> 'b t

val map_es : ('a -> 'b Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a t -> 'b t Tezos_base.TzPervasives.tzresult Lwt.t

val map_ep : ('a -> 'b Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a t -> 'b t Tezos_base.TzPervasives.tzresult Lwt.t